requirements.txt
cycler==0.10.0
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
Pillow==7.2.0
pygame==1.9.6
PyGetWindow==0.0.8
pynput==1.6.8
pyparsing==2.4.7
PyRect==0.1.4
python-dateutil==2.8.1
six==1.15.0